refactor(settings): extract MappingsPage
Also fixes a latent bug the extraction would otherwise have introduced:
the shell's collect() merge is one level deep (config_data[key].update(value)),
which would silently un-delete a removed courier mapping on save. collect()
now mutates the live column_mappings/courier_mappings dicts in place so
deletions actually stick.
